+++ src/main/java/io/ConsoleInputHandler.java
@@ -1,4 +1,4 @@
-package racingcar;
+package io;
 
 import camp.nextstep.edu.missionutils.Console;
