+++ src/main/java/io/ConsoleOutputHandler.java
@@ -1,4 +1,4 @@
-package racingcar;
+package io;
 
 import java.util.List;
 import java.util.stream.IntStream;
