+++ src/main/java/io/OutputHandler.java
@@ -1,4 +1,4 @@
-package racingcar;
+package io;
 
 import java.util.List;
